dailyfresh/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
billiard==3.3.0.23
celery==3.1.24
cffi==1.14.0
cryptography==2.8
Django==1.8
django-celery==3.3.1
django-haystack==2.6.0
django-redis==4.8.0
django-redis-sessions==0.5.6
django-tinymce==2.6.0
importlib-metadata==1.5.0
itsdangerous==1.1.0
jieba==0.42.1
kombu==3.0.37
Pillow==6.2.1
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
pytz==2017.3
redis==2.10.6
six==1.14.0
sqlparse==0.3.1
vine==1.1.4
Whoosh==2.7.4
zipp==2.2.0
